tests/wpt/web-platform-tests/css/work-in-progress/opera/animations/changing_animation_name.html

<!DOCTYPE html>
<html>
<head>
   <title>CSS Animations - changing animation-name during animation</title>
   <link rel="author" title="Opera Software ASA" href="http://www.opera.com/">
   <link rel="help" href="https://drafts.csswg.org/css3-animations/#animations">
   <meta name="flags" content="dom">
   <meta name="assert" content="Changing the value of 'animation-name' does not
   necessarily restart an animation">
   <script src="/resources/testharness.js"></script>
   <script src="/resources/testharnessreport.js"></script>
   <style type="text/css">
      @keyframes slide {
         from {
            left: 0px;
         }
         to {
            left: 100px;
         }
      }
      @keyframes black {
         from {
            background-color: black;
         }
         to {
            background-color: black;
         }
      }
      div {
         width: 100px;
         height: 100px;
         background: blue;
         position: relative;
      }
      .both {
         animation: slide 1s forwards, black 1s forwards;
      }
      .onlymove {
         animation: slide 1s forwards;
      }
   </style>
   <script type="text/javascript">
      setup({explicit_done:true});
      onload = function() {
         var testName = document.querySelector('title').firstChild.data;
         var div = document.querySelector('div');
         var log = document.querySelectorAll('p')[1].firstChild;
         var startCounts = { slide: 0, black: 0 };
         var endCounts = { slide: 0, black: 0 };
         var timeout = setTimeout(function() {
            var passed = startCounts['slide'] == 1 && startCounts['black'] == 1
                         && endCounts['slide'] == 1 && endCounts['black'] == 0;
            test(function(){assert_true(passed);}, testName); done();
            log.data = passed ? 'PASS if the animation occurred.' : 'FAIL';
         }, 1500);
         div.addEventListener('animationstart', function(evt) {
            startCounts[evt.animationName]++;
         });
         div.addEventListener('animationend', function(evt) {
            endCounts[evt.animationName]++;
         });
         div.className = 'both';
         setTimeout(function() {
            div.className = 'onlymove';
         }, 500);
      }
   </script>
   <style>iframe{display:none}</style>
</head>
<body>
   <p>The square below should slide right exactly one time, switching color
   while moving.</p>
   <p>Testing...</p>
   <div></div>
</body>
</html>
